Recovery latency: immediate gap repair, burst-aware FEC, cumulative credit, PTO cap; bench deadline fix
(1) Ack-driven repair fires on the first hole (no packet threshold / reorder window
unless reordering seen), matched against covering repairs/re-sends in flight; 0x83
FEC-feedback frame on every ACK (lowest undelivered + 256-bit map) makes 'missing'
exact. Re-sends on lte: 347 blind -> 21 exact. (2) PathEstimator tracks loss-burst
runs; fecRedundancy = lossRate*(1+burstMean/2) + 2.3 sigma; encoder window 32->128;
decoder rotation with 1024-seq overlap. (3) Grant.creditBytes is now the absolute
cumulative limit (8 B) so any later grant supersedes a lost one; slow start via
credit probes; loss no longer shrinks the target. (4) ptoUs: pto, 1.5, 2, then
doubling; cap 2 s. (5) The Linux low-rate 'tail' was the bench's receive deadline
(warm-up not counted): Main.kt waits max(10 s, 50 RTT) after the last real send;
WSL native 2000/2000. Hardening: ioStats sendErrors/gsoFallback/dropped, GSO runs
capped at 64 segs / 65535-48 B with split + per-datagram fallback (cargo test),
IOExceptions never kill rx/timer threads. Sim before->after (adapt, seed 1):
lte p99 247->117 ms, p999 399->136, late 703->0; starlink p99 58->53.
RecoveryTest x2, grant-blackout test. core 71, transport 40, nativeTest 40.

+++ core/src/main/kotlin/aether/core/Frames.kt
@@ -22,10 +22,17 @@ sealed interface Frame {
         }
     }
 
-    /** Receiver-driven credit: bytes the sender may inject on this path. Core of Homa-style CC. */
-    data class Grant(val path: PathId, val creditBytes: Int, val priority: Int) : Frame {
+    /**
+     * Receiver-driven credit (Homa lineage), **cumulative**: [creditBytes] is the absolute credit limit — the total
+     * number of credit-charged bytes (source, repair, re-send and probe datagrams, see `SenderCredit`) the sender may
+     * have sent on this path since it was set up. It is not a delta: a sender keeps `max(limit, creditBytes)`, so a
+     * later grant supersedes a lost one, re-sent and piggybacked grants are idempotent, and reordered grants are
+     * harmless (v0.5 grants were additive deltas and a lost one stalled the sender until the re-send timer).
+     * Wire: `0x03 path(1) creditBytes(8) priority(1)`.
+     */
+    data class Grant(val path: PathId, val creditBytes: Long, val priority: Int) : Frame {
         override fun write(buf: ByteBuffer) {
-            buf.put(0x03).put(path.raw.toByte()).putInt(creditBytes).put(priority.toByte())
+            buf.put(0x03).put(path.raw.toByte()).putLong(creditBytes).put(priority.toByte())
         }
     }
 
@@ -94,7 +101,7 @@ object FrameCodec {
                 }
                 Frame.Ack(p, l, r, ce, ts)
             }
-            0x03 -> Frame.Grant(PathId(buf.get().toInt() and 0xFF), buf.getInt(), buf.get().toInt() and 0xFF)
+            0x03 -> Frame.Grant(PathId(buf.get().toInt() and 0xFF), buf.getLong(), buf.get().toInt() and 0xFF)
             0x04 -> {
                 val wb = buf.getLong(); val wl = buf.getShort().toInt() and 0xFFFF; val s = buf.getInt()
                 val len = buf.getShort().toInt() and 0xFFFF
